[XTRX.jl] Move autogenerated constants to `CSRs.jl`
Also add a CI job to ensure that the CSRs stay in-sync just like we do
for `csr.h`.
