+++ software/XTRX.jl/src/XTRX.jl
@@ -1,995 +1,5 @@
 module XTRX
 
-const CSR_BASE = Clong(0xf0000000)
+include("CSRs.jl")
 
-const CSR_UART_BASE = CSR_BASE + Clong(0x00)
-
-const CSR_UART_RXTX_ADDR = CSR_BASE + Clong(0x00)
-
-const CSR_UART_RXTX_SIZE = 1
-
-const CSR_UART_TXFULL_ADDR = CSR_BASE + Clong(0x04)
-
-const CSR_UART_TXFULL_SIZE = 1
-
-const CSR_UART_RXEMPTY_ADDR = CSR_BASE + Clong(0x08)
-
-const CSR_UART_RXEMPTY_SIZE = 1
-
-const CSR_UART_EV_STATUS_ADDR = CSR_BASE + Clong(0x0c)
-
-const CSR_UART_EV_STATUS_SIZE = 1
-
-const CSR_UART_EV_STATUS_TX_OFFSET = 0
-
-const CSR_UART_EV_STATUS_TX_SIZE = 1
-
-const CSR_UART_EV_STATUS_RX_OFFSET = 1
-
-const CSR_UART_EV_STATUS_RX_SIZE = 1
-
-const CSR_UART_EV_PENDING_ADDR = CSR_BASE + Clong(0x10)
-
-const CSR_UART_EV_PENDING_SIZE = 1
-
-const CSR_UART_EV_PENDING_TX_OFFSET = 0
-
-const CSR_UART_EV_PENDING_TX_SIZE = 1
-
-const CSR_UART_EV_PENDING_RX_OFFSET = 1
-
-const CSR_UART_EV_PENDING_RX_SIZE = 1
-
-const CSR_UART_EV_ENABLE_ADDR = CSR_BASE + Clong(0x14)
-
-const CSR_UART_EV_ENABLE_SIZE = 1
-
-const CSR_UART_EV_ENABLE_TX_OFFSET = 0
-
-const CSR_UART_EV_ENABLE_TX_SIZE = 1
-
-const CSR_UART_EV_ENABLE_RX_OFFSET = 1
-
-const CSR_UART_EV_ENABLE_RX_SIZE = 1
-
-const CSR_UART_TXEMPTY_ADDR = CSR_BASE + Clong(0x18)
-
-const CSR_UART_TXEMPTY_SIZE = 1
-
-const CSR_UART_RXFULL_ADDR = CSR_BASE + Clong(0x1c)
-
-const CSR_UART_RXFULL_SIZE = 1
-
-const CSR_UART_XOVER_RXTX_ADDR = CSR_BASE + Clong(0x20)
-
-const CSR_UART_XOVER_RXTX_SIZE = 1
-
-const CSR_UART_XOVER_TXFULL_ADDR = CSR_BASE + Clong(0x24)
-
-const CSR_UART_XOVER_TXFULL_SIZE = 1
-
-const CSR_UART_XOVER_RXEMPTY_ADDR = CSR_BASE + Clong(0x28)
-
-const CSR_UART_XOVER_RXEMPTY_SIZE = 1
-
-const CSR_UART_XOVER_EV_STATUS_ADDR = CSR_BASE + Clong(0x2c)
-
-const CSR_UART_XOVER_EV_STATUS_SIZE = 1
-
-const CSR_UART_XOVER_EV_STATUS_TX_OFFSET = 0
-
-const CSR_UART_XOVER_EV_STATUS_TX_SIZE = 1
-
-const CSR_UART_XOVER_EV_STATUS_RX_OFFSET = 1
-
-const CSR_UART_XOVER_EV_STATUS_RX_SIZE = 1
-
-const CSR_UART_XOVER_EV_PENDING_ADDR = CSR_BASE + Clong(0x30)
-
-const CSR_UART_XOVER_EV_PENDING_SIZE = 1
-
-const CSR_UART_XOVER_EV_PENDING_TX_OFFSET = 0
-
-const CSR_UART_XOVER_EV_PENDING_TX_SIZE = 1
-
-const CSR_UART_XOVER_EV_PENDING_RX_OFFSET = 1
-
-const CSR_UART_XOVER_EV_PENDING_RX_SIZE = 1
-
-const CSR_UART_XOVER_EV_ENABLE_ADDR = CSR_BASE + Clong(0x34)
-
-const CSR_UART_XOVER_EV_ENABLE_SIZE = 1
-
-const CSR_UART_XOVER_EV_ENABLE_TX_OFFSET = 0
-
-const CSR_UART_XOVER_EV_ENABLE_TX_SIZE = 1
-
-const CSR_UART_XOVER_EV_ENABLE_RX_OFFSET = 1
-
-const CSR_UART_XOVER_EV_ENABLE_RX_SIZE = 1
-
-const CSR_UART_XOVER_TXEMPTY_ADDR = CSR_BASE + Clong(0x38)
-
-const CSR_UART_XOVER_TXEMPTY_SIZE = 1
-
-const CSR_UART_XOVER_RXFULL_ADDR = CSR_BASE + Clong(0x3c)
-
-const CSR_UART_XOVER_RXFULL_SIZE = 1
-
-const CSR_ICAP_BASE = CSR_BASE + Clong(0x0800)
-
-const CSR_ICAP_ADDR_ADDR = CSR_BASE + Clong(0x0800)
-
-const CSR_ICAP_ADDR_SIZE = 1
-
-const CSR_ICAP_DATA_ADDR = CSR_BASE + Clong(0x0804)
-
-const CSR_ICAP_DATA_SIZE = 1
-
-const CSR_ICAP_WRITE_ADDR = CSR_BASE + Clong(0x0808)
-
-const CSR_ICAP_WRITE_SIZE = 1
-
-const CSR_ICAP_DONE_ADDR = CSR_BASE + Clong(0x080c)
-
-const CSR_ICAP_DONE_SIZE = 1
-
-const CSR_ICAP_READ_ADDR = CSR_BASE + Clong(0x0810)
-
-const CSR_ICAP_READ_SIZE = 1
-
-const CSR_FLASH_BASE = CSR_BASE + Clong(0x1000)
-
-const CSR_FLASH_SPI_CONTROL_ADDR = CSR_BASE + Clong(0x1000)
-
-const CSR_FLASH_SPI_CONTROL_SIZE = 1
-
-const CSR_FLASH_SPI_CONTROL_START_OFFSET = 0
-
-const CSR_FLASH_SPI_CONTROL_START_SIZE = 1
-
-const CSR_FLASH_SPI_CONTROL_LENGTH_OFFSET = 8
-
-const CSR_FLASH_SPI_CONTROL_LENGTH_SIZE = 8
-
-const CSR_FLASH_SPI_STATUS_ADDR = CSR_BASE + Clong(0x1004)
-
-const CSR_FLASH_SPI_STATUS_SIZE = 1
-
-const CSR_FLASH_SPI_STATUS_DONE_OFFSET = 0
-
-const CSR_FLASH_SPI_STATUS_DONE_SIZE = 1
-
-const CSR_FLASH_SPI_MOSI_ADDR = CSR_BASE + Clong(0x1008)
-
-const CSR_FLASH_SPI_MOSI_SIZE = 2
-
-const CSR_FLASH_SPI_MISO_ADDR = CSR_BASE + Clong(0x1010)
-
-const CSR_FLASH_SPI_MISO_SIZE = 2
-
-const CSR_FLASH_SPI_CS_ADDR = CSR_BASE + Clong(0x1018)
-
-const CSR_FLASH_SPI_CS_SIZE = 1
-
-const CSR_FLASH_SPI_CS_SEL_OFFSET = 0
-
-const CSR_FLASH_SPI_CS_SEL_SIZE = 1
-
-const CSR_FLASH_SPI_CS_MODE_OFFSET = 16
-
-const CSR_FLASH_SPI_CS_MODE_SIZE = 1
-
-const CSR_FLASH_SPI_LOOPBACK_ADDR = CSR_BASE + Clong(0x101c)
-
-const CSR_FLASH_SPI_LOOPBACK_SIZE = 1
-
-const CSR_FLASH_SPI_LOOPBACK_MODE_OFFSET = 0
-
-const CSR_FLASH_SPI_LOOPBACK_MODE_SIZE = 1
-
-const CSR_XADC_BASE = CSR_BASE + Clong(0x1800)
-
-const CSR_XADC_TEMPERATURE_ADDR = CSR_BASE + Clong(0x1800)
-
-const CSR_XADC_TEMPERATURE_SIZE = 1
-
-const CSR_XADC_VCCINT_ADDR = CSR_BASE + Clong(0x1804)
-
-const CSR_XADC_VCCINT_SIZE = 1
-
-const CSR_XADC_VCCAUX_ADDR = CSR_BASE + Clong(0x1808)
-
-const CSR_XADC_VCCAUX_SIZE = 1
-
-const CSR_XADC_VCCBRAM_ADDR = CSR_BASE + Clong(0x180c)
-
-const CSR_XADC_VCCBRAM_SIZE = 1
-
-const CSR_XADC_EOC_ADDR = CSR_BASE + Clong(0x1810)
-
-const CSR_XADC_EOC_SIZE = 1
-
-const CSR_XADC_EOS_ADDR = CSR_BASE + Clong(0x1814)
-
-const CSR_XADC_EOS_SIZE = 1
-
-const CSR_DNA_BASE = CSR_BASE + Clong(0x2000)
-
-const CSR_DNA_ID_ADDR = CSR_BASE + Clong(0x2000)
-
-const CSR_DNA_ID_SIZE = 2
-
-const CSR_AUX_BASE = CSR_BASE + Clong(0x2800)
-
-const CSR_AUX_CONTROL_ADDR = CSR_BASE + Clong(0x2800)
-
-const CSR_AUX_CONTROL_SIZE = 1
-
-const CSR_AUX_CONTROL_IOVCC_SEL_OFFSET = 0
-
-const CSR_AUX_CONTROL_IOVCC_SEL_SIZE = 1
-
-const CSR_AUX_CONTROL_EN_SMSIGIO_OFFSET = 1
-
-const CSR_AUX_CONTROL_EN_SMSIGIO_SIZE = 1
-
-const CSR_AUX_CONTROL_OPTION_OFFSET = 2
-
-const CSR_AUX_CONTROL_OPTION_SIZE = 1
-
-const CSR_AUX_CONTROL_GPIO13_OFFSET = 3
-
-const CSR_AUX_CONTROL_GPIO13_SIZE = 1
-
-const CSR_CTRL_BASE = CSR_BASE + Clong(0x3000)
-
-const CSR_CTRL_RESET_ADDR = CSR_BASE + Clong(0x3000)
-
-const CSR_CTRL_RESET_SIZE = 1
-
-const CSR_CTRL_RESET_SOC_RST_OFFSET = 0
-
-const CSR_CTRL_RESET_SOC_RST_SIZE = 1
-
-const CSR_CTRL_RESET_CPU_RST_OFFSET = 1
-
-const CSR_CTRL_RESET_CPU_RST_SIZE = 1
-
-const CSR_CTRL_SCRATCH_ADDR = CSR_BASE + Clong(0x3004)
-
-const CSR_CTRL_SCRATCH_SIZE = 1
-
-const CSR_CTRL_BUS_ERRORS_ADDR = CSR_BASE + Clong(0x3008)
-
-const CSR_CTRL_BUS_ERRORS_SIZE = 1
-
-const CSR_FLASH_CS_N_BASE = CSR_BASE + Clong(0x3800)
-
-const CSR_FLASH_CS_N_OUT_ADDR = CSR_BASE + Clong(0x3800)
-
-const CSR_FLASH_CS_N_OUT_SIZE = 1
-
-const CSR_IDENTIFIER_MEM_BASE = CSR_BASE + Clong(0x4000)
-
-const CSR_LEDS_BASE = CSR_BASE + Clong(0x4800)
-
-const CSR_LEDS_OUT_ADDR = CSR_BASE + Clong(0x4800)
-
-const CSR_LEDS_OUT_SIZE = 1
-
-const CSR_PCIE_PHY_BASE = CSR_BASE + Clong(0x5000)
-
-const CSR_PCIE_PHY_LINK_STATUS_ADDR = CSR_BASE + Clong(0x5000)
-
-const CSR_PCIE_PHY_LINK_STATUS_SIZE = 1
-
-const CSR_PCIE_PHY_LINK_STATUS_STATUS_OFFSET = 0
-
-const CSR_PCIE_PHY_LINK_STATUS_STATUS_SIZE = 1
-
-const CSR_PCIE_PHY_LINK_STATUS_RATE_OFFSET = 1
-
-const CSR_PCIE_PHY_LINK_STATUS_RATE_SIZE = 1
-
-const CSR_PCIE_PHY_LINK_STATUS_WIDTH_OFFSET = 2
-
-const CSR_PCIE_PHY_LINK_STATUS_WIDTH_SIZE = 2
-
-const CSR_PCIE_PHY_LINK_STATUS_LTSSM_OFFSET = 4
-
-const CSR_PCIE_PHY_LINK_STATUS_LTSSM_SIZE = 6
-
-const CSR_PCIE_PHY_MSI_ENABLE_ADDR = CSR_BASE + Clong(0x5004)
-
-const CSR_PCIE_PHY_MSI_ENABLE_SIZE = 1
-
-const CSR_PCIE_PHY_MSIX_ENABLE_ADDR = CSR_BASE + Clong(0x5008)
-
-const CSR_PCIE_PHY_MSIX_ENABLE_SIZE = 1
-
-const CSR_PCIE_PHY_BUS_MASTER_ENABLE_ADDR = CSR_BASE + Clong(0x500c)
-
-const CSR_PCIE_PHY_BUS_MASTER_ENABLE_SIZE = 1
-
-const CSR_PCIE_PHY_MAX_REQUEST_SIZE_ADDR = CSR_BASE + Clong(0x5010)
-
-const CSR_PCIE_PHY_MAX_REQUEST_SIZE_SIZE = 1
-
-const CSR_PCIE_PHY_MAX_PAYLOAD_SIZE_ADDR = CSR_BASE + Clong(0x5014)
-
-const CSR_PCIE_PHY_MAX_PAYLOAD_SIZE_SIZE = 1
-
-const CSR_PCIE_MSI_BASE = CSR_BASE + Clong(0x5800)
-
-const CSR_PCIE_MSI_ENABLE_ADDR = CSR_BASE + Clong(0x5800)
-
-const CSR_PCIE_MSI_ENABLE_SIZE = 1
-
-const CSR_PCIE_MSI_CLEAR_ADDR = CSR_BASE + Clong(0x5804)
-
-const CSR_PCIE_MSI_CLEAR_SIZE = 1
-
-const CSR_PCIE_MSI_VECTOR_ADDR = CSR_BASE + Clong(0x5808)
-
-const CSR_PCIE_MSI_VECTOR_SIZE = 1
-
-const CSR_PCIE_DMA0_BASE = CSR_BASE + Clong(0x6000)
-
-const CSR_PCIE_DMA0_WRITER_ENABLE_ADDR = CSR_BASE + Clong(0x6000)
-
-const CSR_PCIE_DMA0_WRITER_ENABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_ADDR = CSR_BASE + Clong(0x6004)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_SIZE = 2
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_ADDRESS_LSB_OFFSET = 0
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_ADDRESS_LSB_SIZE = 32
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_LENGTH_OFFSET = 32
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_LENGTH_SIZE = 24
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_IRQ_DISABLE_OFFSET = 56
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_IRQ_DISABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_LAST_DISABLE_OFFSET = 57
-
-const CSR_PCIE_DMA0_WRITER_TABLE_VALUE_LAST_DISABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_WE_ADDR = CSR_BASE + Clong(0x600c)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_WE_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_WE_ADDRESS_MSB_OFFSET = 0
-
-const CSR_PCIE_DMA0_WRITER_TABLE_WE_ADDRESS_MSB_SIZE = 32
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_PROG_N_ADDR = CSR_BASE + Clong(0x6010)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_PROG_N_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_ADDR = CSR_BASE + Clong(0x6014)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_INDEX_OFFSET = 0
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_INDEX_SIZE = 16
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_COUNT_OFFSET = 16
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LOOP_STATUS_COUNT_SIZE = 16
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LEVEL_ADDR = CSR_BASE + Clong(0x6018)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_LEVEL_SIZE = 1
-
-const CSR_PCIE_DMA0_WRITER_TABLE_RESET_ADDR = CSR_BASE + Clong(0x601c)
-
-const CSR_PCIE_DMA0_WRITER_TABLE_RESET_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_ENABLE_ADDR = CSR_BASE + Clong(0x6020)
-
-const CSR_PCIE_DMA0_READER_ENABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_ADDR = CSR_BASE + Clong(0x6024)
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_SIZE = 2
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_ADDRESS_LSB_OFFSET = 0
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_ADDRESS_LSB_SIZE = 32
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_LENGTH_OFFSET = 32
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_LENGTH_SIZE = 24
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_IRQ_DISABLE_OFFSET = 56
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_IRQ_DISABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_LAST_DISABLE_OFFSET = 57
-
-const CSR_PCIE_DMA0_READER_TABLE_VALUE_LAST_DISABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_WE_ADDR = CSR_BASE + Clong(0x602c)
-
-const CSR_PCIE_DMA0_READER_TABLE_WE_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_WE_ADDRESS_MSB_OFFSET = 0
-
-const CSR_PCIE_DMA0_READER_TABLE_WE_ADDRESS_MSB_SIZE = 32
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_PROG_N_ADDR = CSR_BASE + Clong(0x6030)
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_PROG_N_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_ADDR = CSR_BASE + Clong(0x6034)
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_INDEX_OFFSET = 0
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_INDEX_SIZE = 16
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_COUNT_OFFSET = 16
-
-const CSR_PCIE_DMA0_READER_TABLE_LOOP_STATUS_COUNT_SIZE = 16
-
-const CSR_PCIE_DMA0_READER_TABLE_LEVEL_ADDR = CSR_BASE + Clong(0x6038)
-
-const CSR_PCIE_DMA0_READER_TABLE_LEVEL_SIZE = 1
-
-const CSR_PCIE_DMA0_READER_TABLE_RESET_ADDR = CSR_BASE + Clong(0x603c)
-
-const CSR_PCIE_DMA0_READER_TABLE_RESET_SIZE = 1
-
-const CSR_PCIE_DMA0_LOOPBACK_ENABLE_ADDR = CSR_BASE + Clong(0x6040)
-
-const CSR_PCIE_DMA0_LOOPBACK_ENABLE_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_ADDR = CSR_BASE + Clong(0x6044)
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_DEPTH_OFFSET = 0
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_DEPTH_SIZE = 24
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_SCRATCH_OFFSET = 24
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_SCRATCH_SIZE = 4
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_LEVEL_MODE_OFFSET = 31
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_CONTROL_LEVEL_MODE_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_STATUS_ADDR = CSR_BASE + Clong(0x6048)
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_STATUS_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_STATUS_LEVEL_OFFSET = 0
-
-const CSR_PCIE_DMA0_BUFFERING_READER_FIFO_STATUS_LEVEL_SIZE = 24
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_ADDR = CSR_BASE + Clong(0x604c)
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_DEPTH_OFFSET = 0
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_DEPTH_SIZE = 24
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_SCRATCH_OFFSET = 24
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_SCRATCH_SIZE = 4
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_LEVEL_MODE_OFFSET = 31
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_CONTROL_LEVEL_MODE_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_STATUS_ADDR = CSR_BASE + Clong(0x6050)
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_STATUS_SIZE = 1
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_STATUS_LEVEL_OFFSET = 0
-
-const CSR_PCIE_DMA0_BUFFERING_WRITER_FIFO_STATUS_LEVEL_SIZE = 24
-
-const CSR_LEDS2_BASE = CSR_BASE + Clong(0x6800)
-
-const CSR_LEDS2_OUT_ADDR = CSR_BASE + Clong(0x6800)
-
-const CSR_LEDS2_OUT_SIZE = 1
-
-const CSR_TIMER0_BASE = CSR_BASE + Clong(0x7000)
-
-const CSR_TIMER0_LOAD_ADDR = CSR_BASE + Clong(0x7000)
-
-const CSR_TIMER0_LOAD_SIZE = 1
-
-const CSR_TIMER0_RELOAD_ADDR = CSR_BASE + Clong(0x7004)
-
-const CSR_TIMER0_RELOAD_SIZE = 1
-
-const CSR_TIMER0_EN_ADDR = CSR_BASE + Clong(0x7008)
-
-const CSR_TIMER0_EN_SIZE = 1
-
-const CSR_TIMER0_UPDATE_VALUE_ADDR = CSR_BASE + Clong(0x700c)
-
-const CSR_TIMER0_UPDATE_VALUE_SIZE = 1
-
-const CSR_TIMER0_VALUE_ADDR = CSR_BASE + Clong(0x7010)
-
-const CSR_TIMER0_VALUE_SIZE = 1
-
-const CSR_TIMER0_EV_STATUS_ADDR = CSR_BASE + Clong(0x7014)
-
-const CSR_TIMER0_EV_STATUS_SIZE = 1
-
-const CSR_TIMER0_EV_STATUS_ZERO_OFFSET = 0
-
-const CSR_TIMER0_EV_STATUS_ZERO_SIZE = 1
-
-const CSR_TIMER0_EV_PENDING_ADDR = CSR_BASE + Clong(0x7018)
-
-const CSR_TIMER0_EV_PENDING_SIZE = 1
-
-const CSR_TIMER0_EV_PENDING_ZERO_OFFSET = 0
-
-const CSR_TIMER0_EV_PENDING_ZERO_SIZE = 1
-
-const CSR_TIMER0_EV_ENABLE_ADDR = CSR_BASE + Clong(0x701c)
-
-const CSR_TIMER0_EV_ENABLE_SIZE = 1
-
-const CSR_TIMER0_EV_ENABLE_ZERO_OFFSET = 0
-
-const CSR_TIMER0_EV_ENABLE_ZERO_SIZE = 1
-
-const CSR_I2C0_BASE = CSR_BASE + Clong(0xa000)
-
-const CSR_I2C0_W_ADDR = CSR_BASE + Clong(0xa000)
-
-const CSR_I2C0_W_SIZE = 1
-
-const CSR_I2C0_W_SCL_OFFSET = 0
-
-const CSR_I2C0_W_SCL_SIZE = 1
-
-const CSR_I2C0_W_OE_OFFSET = 1
-
-const CSR_I2C0_W_OE_SIZE = 1
-
-const CSR_I2C0_W_SDA_OFFSET = 2
-
-const CSR_I2C0_W_SDA_SIZE = 1
-
-const CSR_I2C0_R_ADDR = CSR_BASE + Clong(0xa004)
-
-const CSR_I2C0_R_SIZE = 1
-
-const CSR_I2C0_R_SDA_OFFSET = 0
-
-const CSR_I2C0_R_SDA_SIZE = 1
-
-const CSR_I2C1_BASE = CSR_BASE + Clong(0xa800)
-
-const CSR_I2C1_W_ADDR = CSR_BASE + Clong(0xa800)
-
-const CSR_I2C1_W_SIZE = 1
-
-const CSR_I2C1_W_SCL_OFFSET = 0
-
-const CSR_I2C1_W_SCL_SIZE = 1
-
-const CSR_I2C1_W_OE_OFFSET = 1
-
-const CSR_I2C1_W_OE_SIZE = 1
-
-const CSR_I2C1_W_SDA_OFFSET = 2
-
-const CSR_I2C1_W_SDA_SIZE = 1
-
-const CSR_I2C1_R_ADDR = CSR_BASE + Clong(0xa804)
-
-const CSR_I2C1_R_SIZE = 1
-
-const CSR_I2C1_R_SDA_OFFSET = 0
-
-const CSR_I2C1_R_SDA_SIZE = 1
-
-const CSR_GPIO_BASE = CSR_BASE + Clong(0xb000)
-
-const CSR_GPIO_CONTROL_ADDR = CSR_BASE + Clong(0xb000)
-
-const CSR_GPIO_CONTROL_SIZE = 1
-
-const CSR_GPIO_CONTROL_GPIO_OFFSET = 0
-
-const CSR_GPIO_CONTROL_GPIO_SIZE = 7
-
-const CSR_GPS_BASE = CSR_BASE + Clong(0xb800)
-
-const CSR_GPS_CONTROL_ADDR = CSR_BASE + Clong(0xb800)
-
-const CSR_GPS_CONTROL_SIZE = 1
-
-const CSR_GPS_CONTROL_ENABLE_OFFSET = 0
-
-const CSR_GPS_CONTROL_ENABLE_SIZE = 1
-
-const CSR_GPS_UART_RXTX_ADDR = CSR_BASE + Clong(0xb804)
-
-const CSR_GPS_UART_RXTX_SIZE = 1
-
-const CSR_GPS_UART_TXFULL_ADDR = CSR_BASE + Clong(0xb808)
-
-const CSR_GPS_UART_TXFULL_SIZE = 1
-
-const CSR_GPS_UART_RXEMPTY_ADDR = CSR_BASE + Clong(0xb80c)
-
-const CSR_GPS_UART_RXEMPTY_SIZE = 1
-
-const CSR_GPS_UART_EV_STATUS_ADDR = CSR_BASE + Clong(0xb810)
-
-const CSR_GPS_UART_EV_STATUS_SIZE = 1
-
-const CSR_GPS_UART_EV_STATUS_TX_OFFSET = 0
-
-const CSR_GPS_UART_EV_STATUS_TX_SIZE = 1
-
-const CSR_GPS_UART_EV_STATUS_RX_OFFSET = 1
-
-const CSR_GPS_UART_EV_STATUS_RX_SIZE = 1
-
-const CSR_GPS_UART_EV_PENDING_ADDR = CSR_BASE + Clong(0xb814)
-
-const CSR_GPS_UART_EV_PENDING_SIZE = 1
-
-const CSR_GPS_UART_EV_PENDING_TX_OFFSET = 0
-
-const CSR_GPS_UART_EV_PENDING_TX_SIZE = 1
-
-const CSR_GPS_UART_EV_PENDING_RX_OFFSET = 1
-
-const CSR_GPS_UART_EV_PENDING_RX_SIZE = 1
-
-const CSR_GPS_UART_EV_ENABLE_ADDR = CSR_BASE + Clong(0xb818)
-
-const CSR_GPS_UART_EV_ENABLE_SIZE = 1
-
-const CSR_GPS_UART_EV_ENABLE_TX_OFFSET = 0
-
-const CSR_GPS_UART_EV_ENABLE_TX_SIZE = 1
-
-const CSR_GPS_UART_EV_ENABLE_RX_OFFSET = 1
-
-const CSR_GPS_UART_EV_ENABLE_RX_SIZE = 1
-
-const CSR_GPS_UART_TXEMPTY_ADDR = CSR_BASE + Clong(0xb81c)
-
-const CSR_GPS_UART_TXEMPTY_SIZE = 1
-
-const CSR_GPS_UART_RXFULL_ADDR = CSR_BASE + Clong(0xb820)
-
-const CSR_GPS_UART_RXFULL_SIZE = 1
-
-const CSR_VCTCXO_BASE = CSR_BASE + Clong(0xc000)
-
-const CSR_VCTCXO_CONTROL_ADDR = CSR_BASE + Clong(0xc000)
-
-const CSR_VCTCXO_CONTROL_SIZE = 1
-
-const CSR_VCTCXO_CONTROL_SEL_OFFSET = 0
-
-const CSR_VCTCXO_CONTROL_SEL_SIZE = 1
-
-const CSR_VCTCXO_CONTROL_EN_OFFSET = 1
-
-const CSR_VCTCXO_CONTROL_EN_SIZE = 1
-
-const CSR_VCTCXO_CYCLES_LATCH_ADDR = CSR_BASE + Clong(0xc004)
-
-const CSR_VCTCXO_CYCLES_LATCH_SIZE = 1
-
-const CSR_VCTCXO_CYCLES_ADDR = CSR_BASE + Clong(0xc008)
-
-const CSR_VCTCXO_CYCLES_SIZE = 1
-
-const CSR_RF_SWITCHES_BASE = CSR_BASE + Clong(0xc800)
-
-const CSR_RF_SWITCHES_TX_ADDR = CSR_BASE + Clong(0xc800)
-
-const CSR_RF_SWITCHES_TX_SIZE = 1
-
-const CSR_RF_SWITCHES_RX_ADDR = CSR_BASE + Clong(0xc804)
-
-const CSR_RF_SWITCHES_RX_SIZE = 1
-
-const CSR_LMS7002M_BASE = CSR_BASE + Clong(0xd000)
-
-const CSR_LMS7002M_CONTROL_ADDR = CSR_BASE + Clong(0xd000)
-
-const CSR_LMS7002M_CONTROL_SIZE = 1
-
-const CSR_LMS7002M_CONTROL_RESET_OFFSET = 0
-
-const CSR_LMS7002M_CONTROL_RESET_SIZE = 1
-
-const CSR_LMS7002M_CONTROL_POWER_DOWN_OFFSET = 1
-
-const CSR_LMS7002M_CONTROL_POWER_DOWN_SIZE = 1
-
-const CSR_LMS7002M_CONTROL_TX_ENABLE_OFFSET = 8
-
-const CSR_LMS7002M_CONTROL_TX_ENABLE_SIZE = 1
-
-const CSR_LMS7002M_CONTROL_RX_ENABLE_OFFSET = 9
-
-const CSR_LMS7002M_CONTROL_RX_ENABLE_SIZE = 1
-
-const CSR_LMS7002M_CONTROL_TX_RX_LOOPBACK_ENABLE_OFFSET = 16
-
-const CSR_LMS7002M_CONTROL_TX_RX_LOOPBACK_ENABLE_SIZE = 1
-
-const CSR_LMS7002M_STATUS_ADDR = CSR_BASE + Clong(0xd004)
-
-const CSR_LMS7002M_STATUS_SIZE = 1
-
-const CSR_LMS7002M_STATUS_RX_CLK_ACTIVE_OFFSET = 0
-
-const CSR_LMS7002M_STATUS_RX_CLK_ACTIVE_SIZE = 1
-
-const CSR_LMS7002M_STATUS_RX_FRAME_ACTIVE_OFFSET = 1
-
-const CSR_LMS7002M_STATUS_RX_FRAME_ACTIVE_SIZE = 1
-
-const CSR_LMS7002M_STATUS_RX_FRAME_ALIGNED_OFFSET = 2
-
-const CSR_LMS7002M_STATUS_RX_FRAME_ALIGNED_SIZE = 1
-
-const CSR_LMS7002M_DELAY_ADDR = CSR_BASE + Clong(0xd008)
-
-const CSR_LMS7002M_DELAY_SIZE = 1
-
-const CSR_LMS7002M_DELAY_TX_DELAY_OFFSET = 0
-
-const CSR_LMS7002M_DELAY_TX_DELAY_SIZE = 5
-
-const CSR_LMS7002M_DELAY_RX_DELAY_OFFSET = 8
-
-const CSR_LMS7002M_DELAY_RX_DELAY_SIZE = 5
-
-const CSR_LMS7002M_SPI_CONTROL_ADDR = CSR_BASE + Clong(0xd00c)
-
-const CSR_LMS7002M_SPI_CONTROL_SIZE = 1
-
-const CSR_LMS7002M_SPI_CONTROL_START_OFFSET = 0
-
-const CSR_LMS7002M_SPI_CONTROL_START_SIZE = 1
-
-const CSR_LMS7002M_SPI_CONTROL_LENGTH_OFFSET = 8
-
-const CSR_LMS7002M_SPI_CONTROL_LENGTH_SIZE = 8
-
-const CSR_LMS7002M_SPI_STATUS_ADDR = CSR_BASE + Clong(0xd010)
-
-const CSR_LMS7002M_SPI_STATUS_SIZE = 1
-
-const CSR_LMS7002M_SPI_STATUS_DONE_OFFSET = 0
-
-const CSR_LMS7002M_SPI_STATUS_DONE_SIZE = 1
-
-const CSR_LMS7002M_SPI_MOSI_ADDR = CSR_BASE + Clong(0xd014)
-
-const CSR_LMS7002M_SPI_MOSI_SIZE = 1
-
-const CSR_LMS7002M_SPI_MISO_ADDR = CSR_BASE + Clong(0xd018)
-
-const CSR_LMS7002M_SPI_MISO_SIZE = 1
-
-const CSR_LMS7002M_SPI_CS_ADDR = CSR_BASE + Clong(0xd01c)
-
-const CSR_LMS7002M_SPI_CS_SIZE = 1
-
-const CSR_LMS7002M_SPI_CS_SEL_OFFSET = 0
-
-const CSR_LMS7002M_SPI_CS_SEL_SIZE = 1
-
-const CSR_LMS7002M_SPI_CS_MODE_OFFSET = 16
-
-const CSR_LMS7002M_SPI_CS_MODE_SIZE = 1
-
-const CSR_LMS7002M_SPI_LOOPBACK_ADDR = CSR_BASE + Clong(0xd020)
-
-const CSR_LMS7002M_SPI_LOOPBACK_SIZE = 1
-
-const CSR_LMS7002M_SPI_LOOPBACK_MODE_OFFSET = 0
-
-const CSR_LMS7002M_SPI_LOOPBACK_MODE_SIZE = 1
-
-const CSR_LMS7002M_TX_PATTERN_CONTROL_ADDR = CSR_BASE + Clong(0xd024)
-
-const CSR_LMS7002M_TX_PATTERN_CONTROL_SIZE = 1
-
-const CSR_LMS7002M_TX_PATTERN_CONTROL_ENABLE_OFFSET = 0
-
-const CSR_LMS7002M_TX_PATTERN_CONTROL_ENABLE_SIZE = 1
-
-const CSR_LMS7002M_RX_PATTERN_CONTROL_ADDR = CSR_BASE + Clong(0xd028)
-
-const CSR_LMS7002M_RX_PATTERN_CONTROL_SIZE = 1
-
-const CSR_LMS7002M_RX_PATTERN_CONTROL_ENABLE_OFFSET = 0
-
-const CSR_LMS7002M_RX_PATTERN_CONTROL_ENABLE_SIZE = 1
-
-const CSR_LMS7002M_RX_PATTERN_ERRORS_ADDR = CSR_BASE + Clong(0xd02c)
-
-const CSR_LMS7002M_RX_PATTERN_ERRORS_SIZE = 1
-
-const CONFIG_CLOCK_FREQUENCY = 125000000
-
-const CONFIG_CPU_RESET_ADDR = 0
-
-const CONFIG_CPU_HUMAN_NAME = "VexRiscv_Min"
-
-const CONFIG_CPU_NOP = "nop"
-
-const CONFIG_MAIN_RAM_INIT = 1
-
-const ROM_BOOT_ADDRESS = 1073741824
-
-const DMA_CHANNELS = 1
-
-const DMA_ADDR_WIDTH = 64
-
-const PCIE_DMA0_READER_INTERRUPT = 0
-
-const PCIE_DMA0_WRITER_INTERRUPT = 1
-
-const CONFIG_CSR_DATA_WIDTH = 32
-
-const CONFIG_CSR_ALIGNMENT = 32
-
-const CONFIG_BUS_STANDARD = "WISHBONE"
-
-const CONFIG_BUS_DATA_WIDTH = 32
-
-const CONFIG_BUS_ADDRESS_WIDTH = 32
-
-const CONFIG_BUS_BURSTING = 0
-
-const TIMER0_INTERRUPT = 1
-
-const UART_INTERRUPT = 0
-
-const ROM_BASE = Clong(0x00000000)
-
-const ROM_SIZE = 0x00008000
-
-const SRAM_BASE = Clong(0x10000000)
-
-const SRAM_SIZE = 0x00002000
-
-const MAIN_RAM_BASE = Clong(0x40000000)
-
-const MAIN_RAM_SIZE = 0x00004000
-
-const CSR_SIZE = 0x00010000
-
-const MEM_REGIONS = "ROM       0x00000000 0x8000 \nSRAM      0x10000000 0x2000 \nMAIN_RAM  0x40000000 0x4000 \nCSR       0xf0000000 0x10000 "
-
-const PCIE_FPGA_VENDOR_ID = 0x10ee
-
-const PCIE_FPGA_DEVICE_ID_S7_GEN2_X1 = 0x7021
-
-const PCIE_FPGA_DEVICE_ID_S7_GEN2_X2 = 0x7022
-
-const PCIE_FPGA_DEVICE_ID_S7_GEN2_X4 = 0x7024
-
-const PCIE_FPGA_DEVICE_ID_S7_GEN2_X8 = 0x7028
-
-const PCIE_FPGA_DEVICE_ID_US_GEN2_X1 = 0x8021
-
-const PCIE_FPGA_DEVICE_ID_US_GEN2_X2 = 0x8022
-
-const PCIE_FPGA_DEVICE_ID_US_GEN2_X4 = 0x8024
-
-const PCIE_FPGA_DEVICE_ID_US_GEN2_X8 = 0x8028
-
-const PCIE_FPGA_DEVICE_ID_US_GEN3_X1 = 0x8031
-
-const PCIE_FPGA_DEVICE_ID_US_GEN3_X2 = 0x8032
-
-const PCIE_FPGA_DEVICE_ID_US_GEN3_X4 = 0x8034
-
-const PCIE_FPGA_DEVICE_ID_US_GEN3_X8 = 0x8038
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN2_X1 = 0x9021
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN2_X2 = 0x9022
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN2_X4 = 0x9024
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN2_X8 = 0x9028
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN2_X16 = 0x902f
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN3_X1 = 0x9031
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN3_X2 = 0x9032
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN3_X4 = 0x9034
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN3_X8 = 0x9038
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN3_X16 = 0x903f
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN4_X1 = 0x9041
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN4_X2 = 0x9042
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN4_X4 = 0x9044
-
-const PCIE_FPGA_DEVICE_ID_USP_GEN4_X8 = 0x9048
-
-const DMA_IRQ_DISABLE = 1 << 24
-
-const DMA_LAST_DISABLE = 1 << 25
-
-const DMA_CHANNEL_COUNT = DMA_CHANNELS
-
-const DMA_BUFFER_PER_IRQ = 32
-
-const DMA_BUFFER_COUNT = 256
-
-const DMA_BUFFER_SIZE = 65536
-
-const DMA_BUFFER_TOTAL_SIZE = DMA_BUFFER_COUNT * DMA_BUFFER_SIZE
-
-const PCIE_DMA_WRITER_ENABLE_OFFSET = 0x0000
-
-const PCIE_DMA_WRITER_TABLE_VALUE_OFFSET = 0x0004
-
-const PCIE_DMA_WRITER_TABLE_WE_OFFSET = 0x000c
-
-const PCIE_DMA_WRITER_TABLE_LOOP_PROG_N_OFFSET = 0x0010
-
-const PCIE_DMA_WRITER_TABLE_LOOP_STATUS_OFFSET = 0x0014
-
-const PCIE_DMA_WRITER_TABLE_LEVEL_OFFSET = 0x0018
-
-const PCIE_DMA_WRITER_TABLE_FLUSH_OFFSET = 0x001c
-
-const PCIE_DMA_READER_ENABLE_OFFSET = 0x0020
-
-const PCIE_DMA_READER_TABLE_VALUE_OFFSET = 0x0024
-
-const PCIE_DMA_READER_TABLE_WE_OFFSET = 0x002c
-
-const PCIE_DMA_READER_TABLE_LOOP_PROG_N_OFFSET = 0x0030
-
-const PCIE_DMA_READER_TABLE_LOOP_STATUS_OFFSET = 0x0034
-
-const PCIE_DMA_READER_TABLE_LEVEL_OFFSET = 0x0038
-
-const PCIE_DMA_READER_TABLE_FLUSH_OFFSET = 0x003c
-
-const PCIE_DMA_LOOPBACK_ENABLE_OFFSET = 0x0040
-
-const PCIE_DMA_BUFFERING_READER_FIFO_DEPTH_ADDR = 0x0044
-
-const PCIE_DMA_BUFFERING_READER_FIFO_LEVEL_ADDR = 0x0048
-
-const PCIE_DMA_BUFFERING_WRITER_FIFO_DEPTH_ADDR = 0x004c
-
-const PCIE_DMA_BUFFERING_WRITER_FIFO_LEVEL_ADDR = 0x0050
-
-end # module
+end # module XTRX
